move deploy.js body to hardhat-router main code
this will prevent confusion and drift between deployment scripts for hardhat-router
